# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health/live": {
      "get": {
        "operationId": "HealthController_live",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Process liveness"
          }
        },
        "tags": ["foundation"]
      }
    },
    "/health": {
      "get": {
        "operationId": "HealthController_health",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Application liveness and dependency status"
          }
        },
        "tags": ["foundation"]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "minLength": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated session"
          }
        },
        "tags": ["identity"]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["identity"]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["identity"]
      }
    },
    "/organizations/{organizationId}/members": {
      "get": {
        "operationId": "TenancyController_members",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["tenancy"]
      }
    },
    "/catalog/categories": {
      "get": {
        "operationId": "CatalogController_categories",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["catalog"]
      }
    },
    "/catalog/categories/{slug}": {
      "get": {
        "operationId": "CatalogController_category",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["catalog"]
      }
    },
    "/catalog/facets": {
      "get": {
        "operationId": "CatalogController_facets",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["catalog"]
      }
    },
    "/catalog/sitemap": {
      "get": {
        "operationId": "CatalogController_sitemap",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["catalog"]
      }
    },
    "/catalog/products": {
      "get": {
        "operationId": "CatalogController_products",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["catalog"]
      }
    },
    "/organizations/{organizationId}/catalog/products/{productId}/archive": {
      "post": {
        "operationId": "CatalogController_archive",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["catalog"]
      }
    },
    "/catalog/products/{slug}": {
      "get": {
        "operationId": "CatalogController_product",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["catalog"]
      }
    },
    "/organizations/{organizationId}/catalog/products": {
      "get": {
        "operationId": "CatalogController_sellerProducts",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["catalog"]
      },
      "post": {
        "operationId": "CatalogController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["catalog"]
      }
    },
    "/organizations/{organizationId}/catalog/categories": {
      "get": {
        "operationId": "CatalogController_sellerCategories",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["catalog"]
      }
    },
    "/organizations/{organizationId}/catalog/products/{productId}": {
      "get": {
        "operationId": "CatalogController_sellerProduct",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["catalog"]
      },
      "patch": {
        "operationId": "CatalogController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["catalog"]
      }
    },
    "/organizations/{organizationId}/catalog/products/{productId}/submit": {
      "post": {
        "operationId": "CatalogController_submit",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["catalog"]
      }
    },
    "/organizations/{organizationId}/catalog/products/{productId}/moderation": {
      "get": {
        "operationId": "CatalogController_moderationHistory",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["catalog"]
      },
      "post": {
        "operationId": "CatalogController_moderate",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["catalog"]
      }
    },
    "/admin/product-moderation": {
      "get": {
        "operationId": "CatalogController_moderationQueue",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["catalog"]
      }
    },
    "/admin/product-moderation/{reviewId}/comments": {
      "post": {
        "operationId": "CatalogController_addModerationComment",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["catalog"]
      }
    },
    "/organizations/{organizationId}/imports/shopify": {
      "post": {
        "operationId": "ImportController_shopify",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["imports"]
      }
    },
    "/organizations/{organizationId}/imports/web": {
      "post": {
        "operationId": "ImportController_web",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["imports"]
      }
    },
    "/organizations/{organizationId}/imports/{jobId}": {
      "get": {
        "operationId": "ImportController_report",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["imports"]
      }
    },
    "/organizations/{organizationId}/imports/{jobId}/retry": {
      "post": {
        "operationId": "ImportController_retry",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["imports"]
      }
    },
    "/organizations/{organizationId}/imports/{jobId}/apply": {
      "post": {
        "operationId": "ImportController_apply",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["imports"]
      }
    },
    "/organizations/{organizationId}/imports/catalog/export.csv": {
      "get": {
        "operationId": "ImportController_exportCsv",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["imports"]
      }
    },
    "/organizations/{organizationId}/catalog/products/{productId}/media/upload-url": {
      "post": {
        "operationId": "MediaController_uploadUrl",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["media"]
      }
    },
    "/organizations/{organizationId}/catalog/products/{productId}/media/{mediaId}/complete": {
      "post": {
        "operationId": "MediaController_complete",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["media"]
      }
    },
    "/organizations/{organizationId}/catalog/products/{productId}/media/{mediaId}": {
      "patch": {
        "operationId": "MediaController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["media"]
      }
    },
    "/storefronts/resolve": {
      "get": {
        "operationId": "StorefrontController_resolve",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["storefronts"]
      }
    },
    "/storefronts/{slug}": {
      "get": {
        "operationId": "StorefrontController_home",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["storefronts"]
      }
    },
    "/storefronts/{slug}/products/{productSlug}": {
      "get": {
        "operationId": "StorefrontController_product",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["storefronts"]
      }
    },
    "/storefronts/{slug}/policies/{policySlug}": {
      "get": {
        "operationId": "StorefrontController_policy",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["storefronts"]
      }
    },
    "/storefronts/{slug}/redirect": {
      "get": {
        "operationId": "StorefrontController_redirect",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["storefronts"]
      }
    },
    "/dealer-applications/mine": {
      "get": {
        "operationId": "DealerController_mine",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["dealers"]
      }
    },
    "/dealer-applications": {
      "post": {
        "operationId": "DealerController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["dealers"]
      }
    },
    "/organizations/{organizationId}/dealer-application": {
      "get": {
        "operationId": "DealerController_owned",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["dealers"]
      },
      "patch": {
        "operationId": "DealerController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["dealers"]
      }
    },
    "/organizations/{organizationId}/dealer-application/submit": {
      "post": {
        "operationId": "DealerController_submit",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["dealers"]
      }
    },
    "/admin/dealer-applications": {
      "get": {
        "operationId": "DealerController_queue",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["dealers"]
      }
    },
    "/admin/dealer-applications/{applicationId}": {
      "get": {
        "operationId": "DealerController_detail",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["dealers"]
      }
    },
    "/admin/dealer-applications/{applicationId}/review": {
      "post": {
        "operationId": "DealerController_review",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["dealers"]
      }
    },
    "/notifications": {
      "get": {
        "operationId": "NotificationController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["notifications"]
      }
    },
    "/notifications/{notificationId}/read": {
      "patch": {
        "operationId": "NotificationController_markRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["notifications"]
      }
    }
  },
  "info": {
    "title": "Project Atlas Foundation API",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "atlas_session"
      }
    },
    "schemas": {}
  }
}